requirements.dev.txt
black>=23.1.0,<23.2
pre-commit>=3.1.1,<3.2
pytest-django>=4.5.2,<4.6
model-bakery>=1.10.1,<1.11
